requirements.txt
customtkinter
pillow
pymupdf
